# specs/004-map-zoom-annotations/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MolfMaps API
  version: 0.3.0
  description: >
    REST contract after adding map annotations (feature 004-map-zoom-annotations).
    Only the annotation endpoints are new. The endpoints inherited from 001/002/003
    (/api/map, /api/login, /api/logout, /api/session, /api/maps*) are unchanged and
    are documented in specs/003-maps-list-admin/contracts/openapi.yaml; they are
    referenced but not reproduced here. The authoritative runtime schema remains the
    OpenAPI document generated by FastAPI.

    Authorization model: listing annotations and fetching point-of-interest images
    is public — viewing maps, following text links, and opening popups never require
    a session (FR-048). Creating, editing, resizing, repositioning, and deleting
    annotations, and attaching or detaching images, all require a valid admin session
    and are enforced server-side (FR-045, FR-046). Hiding the placement toggles in the
    UI is presentation only.

    Zoom and pan are entirely client-side view state and have no API surface
    (research Decision 1). Placement mode is likewise client-only and is never sent
    to the server.
servers:
  - url: http://localhost:8000
    description: Local backend (Uvicorn). Angular dev uses a proxy so the browser
      calls same-origin /api/* on the ng serve origin.
paths:
...
  /api/maps/{map_id}:
    parameters:
      - $ref: "#/components/parameters/MapId"
    get:
      summary: Get one map's metadata (unchanged from 003)
      description: >
        Reproduced here only because following a text link resolves the target
        through this endpoint. A target map deleted after the link was saved
        answers 404 with "This map is no longer available.", which is what the
        frontend shows instead of a broken view (FR-030). The contract itself is
        unchanged.
      operationId: getMapById
      security: []
      responses:
        "200":
          description: The requested map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapSummary"
        "404":
          description: No map with this id exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a map (admin only, behavior extended)
      description: >
        Unchanged request and response contract from 003. Behavior is extended:
        deleting a map now also removes its annotations and unlinks their
        point-of-interest image files (FR-044). Text links on *other* maps that
        pointed at this one are deliberately left in place and simply report
        `target_available: false` afterwards (FR-030).
      operationId: deleteMap
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Map, its image, and its annotations deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: No valid admin session. The map remains available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No map with this id exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    MapId:
      name: map_id
      in: path
      required: true
      description: >
        Server-generated map identifier (32 lowercase hex characters). Declared
        as an opaque string at runtime, so an id of any other shape resolves to
        404 rather than a validation error.
      schema:
        type: string
        pattern: "^[0-9a-f]{32}$"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    MapSummary:
      type: object
      required: [id, name, image_url]
      description: Unchanged from 003; referenced when following a text link.
      properties:
        id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        name:
          type: string
          minLength: 1
          maxLength: 100
        image_url:
          type: string
    MessageResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: >
            Generic message safe to show a user. Never reveals internal paths or
            stack details.